generate-policy-bot.yml: Add script to generate Policy Bot config
This is a script to maintain a [Policy Bot][policy-bot] configuration.
It will look at all the workflows and generate a config which will
enforce they all pass.

If the workflow has a path filter then this will be added to the policy
bot config too. In this way, we implement "requires if triggered"
workflows, which GitHub does not support itself.

Another workflow can be supplied too. If it was, it will be merged with
the generated one. In that way, we can have a local config alongside a
generated one.

[policy-bot]: https://github.com/palantir/policy-bot/