generate-policy-bot.yml: Enforce workflows on the right target branches
We're currently not considering branch filters in our policy. That means
that PRs which target branches other than `master` (e.g. stacked PRs)
are not able to be merged because they're waiting for a run that won't
happen.

Fix this by adding a `TargetBranches` condition if a branch filter is
present. This only activates the check for a workflow run if the PR
targets one of the relevant branches.

In the YAML schema for policy bot, the type of this is slightly
different to the type for changed paths. It's a single regexp rather
than a list. So we need to make alternates for each case and then
combine them with `|`.