policy bot: Add an override command
There's a gap at the minute which it's not clear how to solve. Policies
are evaluated against the target branch. We require workflows to have
successful runs if their triggers are touched. This means it's hard to
delete a workflow if it triggers when its own file is modified, which is
common, since the policy on `master` will expect a run but one won't
happen.

To help with this and potentially any other cases, here we're adding the
capability to override the bot by typing a comment `policy bot: approve`
in the PR.

A bit of work was needed to integrate this. We've got support for
merging a static config with our generated one, but it couldn't insert
new policies in a way that they would be alternatives to the GitHub
Actions policies which we generate. Here we're adding some support for
doing that, and moving the generated policies under an `and` block so
that the override can be added alongside.

Co-authored-by: Dimitris Sotirakis <dimitrios.sotirakis@grafana.com>