fix(ci): drop local validate rule
We have a local rule to validate that `.policy.yml` itself is correct
when modified. It's requiring a status, but we don't have a workflow
producing it right now. I'll bring this back in an upcoming PR, but for
the meantime I need to drop this so I can merge other PRs which touch
the config.