chore(release): make releases with with github actions user
Since the GitHub Actions user creates our release tag and release, it
doesn't trigger the actual build. But we can do the same thing we do for
the release PR: create the release in draft mode, and then a human
publishes. Since a human takes that action, it should trigger the build.

Also, don't include the project name in the tag names. It's not required
since this isn't a monorepo.