fix: generate extended and doublestar globs
At some point we updated to a new version of `go-globre` and this
disabled extended and doublestar glob support by default. They're opted
into by passing a couple of flags. That's what we do here.

This was caught by failing tests, but sadly we weren't actually running
tests in CI. That's fixed now too, by adding a new job before the
`build` job.